Solucionado el problema de compilacion del TP. Ahora no anda el test runner

+++ tp2/src/common_Socket.h
@@ -17,7 +17,7 @@ private:
 	socklen_t ai_addrlen;
 	struct sockaddr *ai_addr;
 	bool conectado;
-	bool enviar(const char *buffer, int tamanio);
+	bool enviar(const char *buffer, ssize_t tamanio);
 	bool recibir(char *buffer, int tamanio);
 
 public:
